Split "flags" structs out of Style::ComputedStyleBase https://bugs.webkit.org/show_bug.cgi?id=311119
Reviewed by NOBODY (OOPS!).

Makes Style::ComputedStyleBase a bit less big by splitting out
the nested "flags" structs.

Remove some unnecessary includes and addressed fallout, including
updating the results of `run-css-property-code-generation-tests`.

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/css/scripts/process-css-properties.py:
* Source/WebCore/css/scripts/test/TestCSSPropertiesResults/StyleComputedStyleProperties.h:
* Source/WebCore/css/scripts/test/TestCSSPropertiesResults/StyleExtractorGenerated.cpp:
* Source/WebCore/rendering/style/RenderStyle.h:
* Source/WebCore/style/StyleChangedAnimatablePropertiesCustom.h:
* Source/WebCore/style/computed/StyleComputedStyleBase+ConstructionInlines.h:
* Source/WebCore/style/computed/StyleComputedStyleBase+GettersInlines.h:
* Source/WebCore/style/computed/StyleComputedStyleBase+SettersInlines.h:
* Source/WebCore/style/computed/StyleComputedStyleBase.cpp:
* Source/WebCore/style/computed/StyleComputedStyleBase.h:
* Source/WebCore/style/computed/data/StyleInheritedFlags.cpp: Added.
* Source/WebCore/style/computed/data/StyleInheritedFlags.h: Added.
* Source/WebCore/style/computed/data/StyleNonInheritedFlags.cpp: Added.
* Source/WebCore/style/computed/data/StyleNonInheritedFlags.h: Added.
* Source/WebCore/style/values/shapes/StyleShapeFunction.cpp:
