+++ Source/WebCore/style/computed/StyleComputedStyleBase.h
@@ -28,15 +28,14 @@
 
 #include <WebCore/BoxExtents.h>
 #include <WebCore/PseudoElementIdentifier.h>
-#include <WebCore/StyleGridAutoFlow.h>
+#include <WebCore/StyleInheritedFlags.h>
+#include <WebCore/StyleNonInheritedFlags.h>
 #include <WebCore/StylePrimitiveNumeric+Forward.h>
-#include <WebCore/WritingMode.h>
 #include <unicode/utypes.h>
 #include <wtf/CheckedRef.h>
 #include <wtf/DataRef.h>
-#include <wtf/FixedVector.h>
+#include <wtf/HashMap.h>
 #include <wtf/OptionSet.h>
-#include <wtf/Vector.h>
 
 namespace WebCore {
 
@@ -269,6 +268,7 @@ struct FontVariationSettings;
 struct FontWeight;
 struct FontWidth;
 struct GapGutter;
+struct GridAutoFlow;
 struct GridPosition;
 struct GridTemplateAreas;
 struct GridTemplateList;
@@ -438,11 +438,6 @@ using WebkitBorderSpacing = Length<CSS::NonnegativeUnzoomed>;
 using WebkitBoxFlex = Number<CSS::All, float>;
 using WebkitBoxFlexGroup = Integer<CSS::Nonnegative>;
 using WebkitBoxOrdinalGroup = Integer<CSS::Positive>;
-
-constexpr auto PublicPseudoIDBits = 19;
-constexpr auto TextDecorationLineBits = 5;
-constexpr auto TextTransformBits = 6;
-constexpr auto PseudoElementTypeBits = 5;
 
 using PseudoStyleCache = HashMap<PseudoElementIdentifier, std::unique_ptr<RenderStyle>>;
 
@@ -721,96 +716,6 @@ public:
     // `@page size`
     inline const PageSize& pageSize() const LIFETIME_BOUND;
     inline void setPageSize(PageSize&&);
-
-    struct NonInheritedFlags {
-        bool operator==(const NonInheritedFlags&) const = default;
-
-        inline void copyNonInheritedFrom(const NonInheritedFlags&);
-
-        inline bool hasAnyPublicPseudoStyles() const;
-        bool hasPseudoStyle(PseudoElementType) const;
-        void setHasPseudoStyles(EnumSet<PseudoElementType>);
-
-#if !LOG_DISABLED
-        void dumpDifferences(TextStream&, const NonInheritedFlags&) const;
-#endif
-
-        PREFERRED_TYPE(Style::DisplayType) unsigned display : 5;
-        PREFERRED_TYPE(Style::DisplayType) unsigned originalDisplay : 5;
-        PREFERRED_TYPE(Overflow) unsigned overflowX : 3;
-        PREFERRED_TYPE(Overflow) unsigned overflowY : 3;
-        PREFERRED_TYPE(Clear) unsigned clear : 3;
-        PREFERRED_TYPE(PositionType) unsigned position : 3;
-        PREFERRED_TYPE(UnicodeBidi) unsigned unicodeBidi : 3;
-        PREFERRED_TYPE(Float) unsigned floating : 3;
-
-        PREFERRED_TYPE(bool) unsigned usesViewportUnits : 1;
-        PREFERRED_TYPE(bool) unsigned usesContainerUnits : 1;
-        PREFERRED_TYPE(bool) unsigned useTreeCountingFunctions : 1;
-        PREFERRED_TYPE(bool) unsigned hasExplicitlyInheritedProperties : 1; // Explicitly inherits a non-inherited property.
-        PREFERRED_TYPE(bool) unsigned disallowsFastPathInheritance : 1;
-
-        // Non-property related state bits.
-        PREFERRED_TYPE(bool) unsigned emptyState : 1;
-        PREFERRED_TYPE(bool) unsigned firstChildState : 1;
-        PREFERRED_TYPE(bool) unsigned lastChildState : 1;
-        PREFERRED_TYPE(bool) unsigned isLink : 1;
-        PREFERRED_TYPE(PseudoElementType) unsigned pseudoElementType : PseudoElementTypeBits;
-        unsigned pseudoBits : PublicPseudoIDBits;
-        unsigned textDecorationLine : TextDecorationLineBits; // Text decorations defined *only* by this element. PREFERRED_TYPE elided to avoid header inclusion.
-
-        // If you add more style bits here, you will also need to update ComputedStyleBase::NonInheritedFlags::copyNonInheritedFrom().
-    };
-
-    struct InheritedFlags {
-        bool operator==(const InheritedFlags&) const = default;
-
-#if !LOG_DISABLED
-        void dumpDifferences(TextStream&, const InheritedFlags&) const;
-#endif
-
-        // Writing Mode = 8 bits (can be packed into 6 if needed)
-        WritingMode writingMode;
-
-        // Text Formatting = 19 bits aligned onto 2 bytes + 4 trailing bits
-        PREFERRED_TYPE(WhiteSpaceCollapse) unsigned char whiteSpaceCollapse : 3;
-        PREFERRED_TYPE(TextWrapMode) unsigned char textWrapMode : 1;
-        PREFERRED_TYPE(TextAlign) unsigned char textAlign : 4;
-        PREFERRED_TYPE(TextWrapStyle) unsigned char textWrapStyle : 2;
-        unsigned char textTransform : TextTransformBits; // PREFERRED_TYPE elided to avoid header inclusion.
-        unsigned char : 1; // byte alignment
-        unsigned char textDecorationLineInEffect : TextDecorationLineBits; // PREFERRED_TYPE elided to avoid header inclusion.
-
-        // Cursors and Visibility = 13 bits aligned onto 4 bits + 1 byte + 1 bit
-        PREFERRED_TYPE(PointerEvents) unsigned char pointerEvents : 4;
-        PREFERRED_TYPE(Visibility) unsigned char visibility : 2;
-        PREFERRED_TYPE(CursorType) unsigned char cursorType : 6;
-#if ENABLE(CURSOR_VISIBILITY)
-        PREFERRED_TYPE(CursorVisibility) unsigned char cursorVisibility : 1;
-#endif
-
-        // Display Type-Specific = 5 bits
-        PREFERRED_TYPE(ListStylePosition) unsigned char listStylePosition : 1;
-        PREFERRED_TYPE(EmptyCell) unsigned char emptyCells : 1;
-        PREFERRED_TYPE(BorderCollapse) unsigned char borderCollapse : 1;
-        PREFERRED_TYPE(CaptionSide) unsigned char captionSide : 2;
-
-        // -webkit- Stuff = 2 bits
-        PREFERRED_TYPE(BoxDirection) unsigned char boxDirection : 1;
-        PREFERRED_TYPE(WebCore::Order) unsigned char rtlOrdering : 1;
-
-        // Color Stuff = 4 bits
-        PREFERRED_TYPE(bool) unsigned char hasExplicitlySetColor : 1;
-        PREFERRED_TYPE(PrintColorAdjust) unsigned char printColorAdjust : 1;
-        PREFERRED_TYPE(InsideLink) unsigned char insideLink : 2;
-
-        PREFERRED_TYPE(bool) unsigned char isZoomed : 1;
-
-#if ENABLE(TEXT_AUTOSIZING)
-        unsigned autosizeStatus : 5;
-#endif
-        // Total = 63 bits (fits in 8 bytes)
-    };
 
 protected:
     friend class Adjuster;
